# be/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sheraton Makkah Digital Sales Experience API
  version: 0.1.0
  description: |
    Single source of truth for the HTTP contract between the Go backend and the
    Next.js frontend. Server stubs and the typed client are generated from this file.

    All error responses use RFC 9457 `application/problem+json`.

    ## Authentication
    Operations are protected by default (global `cookieAuth`); public operations opt out
    explicitly. Login sets two `HttpOnly; SameSite=Strict` cookies: a short-lived access token
    (`dp_access`, `__Host-dp_access` over HTTPS) and a rotating refresh token (`dp_refresh`,
    `__Secure-dp_refresh` over HTTPS) scoped to `/api/v1/auth`. Browsers send them automatically;
    clients never handle tokens.

    ## CSRF
    State-changing requests (`POST`, `PUT`, `PATCH`, `DELETE`) under `/api/v1/auth` and
    `/api/v1/admin` must send `X-CSRF-Protection: 1`, and their `Origin`, when present, must be
    the site origin. Otherwise the API answers `403`.
  license:
    name: Proprietary
    url: https://github.com/omeryilmazbusiness/Digital-profile
servers:
  - url: /
    description: Same origin — the reverse proxy routes API paths to the backend
security:
  - cookieAuth: []
tags:
  - name: system
    description: Operational endpoints (liveness / readiness probes)
  - name: auth
    description: Admin session management
  - name: media
    description: Image library (admin) and immutable image delivery (public)

paths:
  /healthz:
    get:
      tags: [system]
      operationId: getLiveness
      summary: Liveness probe
      description: Reports whether the process is running. Never checks dependencies.
      security: []
      responses:
        "200":
          description: Process is alive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthReport"
        default:
          $ref: "#/components/responses/Problem"

  /readyz:
    get:
      tags: [system]
      operationId: getReadiness
      summary: Readiness probe
      description: Reports whether the service can accept traffic. Runs every registered dependency check.
      security: []
      responses:
        "200":
          description: All dependency checks passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthReport"
        "503":
          description: At least one dependency check failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthReport"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Sign in
      description: |
        Verifies the credentials and starts a session. The response never reveals whether the
        email exists or the account is temporarily locked after repeated failures.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in; session cookies are set
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookies"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshSession
      summary: Rotate the session tokens
      description: |
        Exchanges the refresh cookie for a new access and refresh token. Each refresh token
        works once; presenting a used one again revokes every session of the account.
        `409` means a concurrent request already rotated the token: retry the original
        request with the cookies it set.
      security: []
      responses:
        "200":
          description: Tokens rotated; session cookies are set
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookies"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Sign out
      description: Revokes the current session and clears the cookies. Idempotent.
      security: []
      responses:
        "204":
          description: Signed out; session cookies are cleared
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookies"
        "403":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/auth/me:
    get:
      tags: [auth]
      operationId: getCurrentAdmin
      summary: Current admin
      responses:
        "200":
          description: The signed-in admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/auth/password:
    put:
      tags: [auth]
      operationId: changePassword
      summary: Change password
      description: |
        Changes the password, signs out every other session and starts a fresh one for the
        caller (new cookies are set).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed; new session cookies are set
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookies"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/admin/media:
    get:
      tags: [media]
      operationId: listMedia
      summary: List images
      description: Newest first. Pass `nextCursor` from the previous page as `cursor`.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
        - name: cursor
          in: query
          schema:
            type: string
            maxLength: 128
            pattern: "^[A-Za-z0-9_-]+$"
      responses:
        "200":
          description: A page of images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    post:
      tags: [media]
      operationId: uploadMedia
      summary: Upload an image
      description: |
        Accepts JPEG, PNG or WebP, identified from the file content rather than its name.
        The original is not stored: responsive WebP variants (480, 960 and 1600 px wide, never
        upscaled) are rendered with EXIF orientation applied and all metadata except the colour
        profile removed. Uploading the same file again returns the existing image with `200`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "200":
          description: The same file was uploaded before; the existing image is returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "201":
          description: Image stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "413":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/admin/media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      tags: [media]
      operationId: getMedia
      summary: Get an image
      responses:
        "200":
          description: The image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      tags: [media]
      operationId: deleteMedia
      summary: Delete an image
      description: Fails with `409` while the image is used by any content.
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/admin/media/{mediaId}/alt-text:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    put:
      tags: [media]
      operationId: updateMediaAltText
      summary: Replace the alternative text
      description: Sets the alt text per language; languages left out are cleared.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocalizedAltText"
      responses:
        "200":
          description: Updated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

  /api/v1/public/media/{file}:
    get:
      tags: [media]
      operationId: getPublicMedia
      summary: Image variant
      description: |
        Serves a variant by its content-hash file name. The bytes behind a name never change,
        so responses are cacheable forever (`immutable`).
      security: []
      parameters:
        - name: file
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}\\.webp$"
        - name: If-None-Match
          in: header
          schema:
            type: string
            maxLength: 1024
      responses:
        "200":
          description: The image
          headers:
            Cache-Control:
              schema:
                type: string
              example: public, max-age=31536000, immutable
            ETag:
              schema:
                type: string
          content:
            image/webp:
              schema:
                type: string
                format: binary
        "304":
          description: The client copy is current
          headers:
            Cache-Control:
              schema:
                type: string
            ETag:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"

components:
  parameters:
    MediaId:
      name: mediaId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dp_access
      description: Access token cookie (`__Host-dp_access` when served over HTTPS)

  headers:
    SessionCookies:
      description: Access and refresh token cookies (HttpOnly, SameSite=Strict)
      schema:
        type: string

  responses:
    Problem:
      description: Error response (RFC 9457)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"

  schemas:
    HealthStatus:
      type: string
      enum: [up, down]

    HealthReport:
      type: object
      additionalProperties: false
      required: [status, version]
      properties:
        status:
          $ref: "#/components/schemas/HealthStatus"
        version:
          type: string
          description: Build version of the running binary
          example: 0.1.0
        checks:
          type: array
          items:
            $ref: "#/components/schemas/HealthCheck"

    HealthCheck:
      type: object
      additionalProperties: false
      required: [name, status]
      properties:
        name:
          type: string
          example: database
        status:
          $ref: "#/components/schemas/HealthStatus"
        error:
          type: string
          description: Failure reason; present only when status is `down`

    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status]
      properties:
        type:
          type: string
          description: URI reference identifying the problem type
          example: about:blank
        title:
          type: string
          example: Not Found
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          example: 404
        detail:
          type: string
        instance:
          type: string
        requestId:
          type: string
          description: Correlates the response with server logs
        errors:
          type: array
          description: Field-level validation errors
          items:
            $ref: "#/components/schemas/FieldError"

    FieldError:
      type: object
      additionalProperties: false
      required: [field, message]
      properties:
        field:
          type: string
          example: checkOut
        message:
          type: string
          example: must be after checkIn

    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          example: momen@example.com
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 1024

    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
          minLength: 1
          maxLength: 1024
        newPassword:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          description: 12–128 characters, different from the current password and the email

    AdminUser:
      type: object
      additionalProperties: false
      required: [id, email, passwordChangedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        lastLoginAt:
          type: string
          format: date-time
        passwordChangedAt:
          type: string
          format: date-time

    SessionInfo:
      type: object
      additionalProperties: false
      required: [accessTokenExpiresAt]
      properties:
        accessTokenExpiresAt:
          type: string
          format: date-time
          description: When to refresh; the access cookie stops being sent after this instant

    LoginResponse:
      type: object
      additionalProperties: false
      required: [admin, accessTokenExpiresAt]
      properties:
        admin:
          $ref: "#/components/schemas/AdminUser"
        accessTokenExpiresAt:
          type: string
          format: date-time

    LocalizedAltText:
      type: object
      additionalProperties: false
      description: Alternative text per interface language
      properties:
        en:
          $ref: "#/components/schemas/AltText"
        id:
          $ref: "#/components/schemas/AltText"
        ar:
          $ref: "#/components/schemas/AltText"

    AltText:
      type: string
      minLength: 1
      maxLength: 300

    ImageVariant:
      type: object
      additionalProperties: false
      required: [width, height, url, byteSize]
      properties:
        width:
          type: integer
          format: int32
        height:
          type: integer
          format: int32
        url:
          type: string
          description: Immutable URL; relative to the API origin unless a CDN base is configured
          example: /api/v1/public/media/9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08.webp
        byteSize:
          type: integer
          format: int64

    Media:
      type: object
      additionalProperties: false
      required: [id, width, height, sourceType, sourceBytes, originalFilename, placeholder, altText, variants, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        width:
          type: integer
          format: int32
          description: Display width after EXIF orientation
        height:
          type: integer
          format: int32
        sourceType:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        sourceBytes:
          type: integer
          format: int64
        originalFilename:
          type: string
        placeholder:
          type: string
          description: Tiny blurred WebP data URI to show while the image loads
        altText:
          $ref: "#/components/schemas/LocalizedAltText"
        variants:
          type: array
          description: Ascending by width; use as `srcset`
          minItems: 1
          items:
            $ref: "#/components/schemas/ImageVariant"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    MediaPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        nextCursor:
          type: string
          description: Present when more images exist