# be/internal/modules/media/store/models.gen.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package store

import (
	"net/netip"
	"time"

	"github.com/google/uuid"
)

type AdminUser struct {
	ID                uuid.UUID
	Email             string
	PasswordHash      string
	PasswordChangedAt time.Time
	FailedLoginCount  int32
	LockedUntil       *time.Time
	LastLoginAt       *time.Time
	CreatedAt         time.Time
	UpdatedAt         time.Time
}

type AuthAuditEvent struct {
	ID        int64
	UserID    *uuid.UUID
	Event     string
	Email     *string
	Ip        *netip.Addr
	UserAgent *string
	CreatedAt time.Time
}

type AuthSession struct {
	ID           uuid.UUID
	UserID       uuid.UUID
	CreatedAt    time.Time
	LastSeenAt   time.Time
	ExpiresAt    time.Time
	RevokedAt    *time.Time
	RevokeReason *string
	Ip           *netip.Addr
	UserAgent    *string
}

type Media struct {
	ID               uuid.UUID
	Checksum         []byte
	OriginalFilename string
	SourceType       string
	SourceBytes      int64
	Width            int32
	Height           int32
	Placeholder      string
	CreatedAt        time.Time
	UpdatedAt        time.Time
}

type MediaTranslation struct {
	MediaID uuid.UUID
	Locale  string
	AltText string
}

type MediaVariant struct {
	MediaID     uuid.UUID
	Width       int32
	Height      int32
	StorageKey  string
	ContentType string
	ByteSize    int64
}

type RefreshToken struct {
	ID        uuid.UUID
	SessionID uuid.UUID
	TokenHash []byte
	ExpiresAt time.Time
	UsedAt    *time.Time
	CreatedAt time.Time
}